testing with adapter

+++ esp8266-freertos/main/power_save.h
@@ -3,7 +3,7 @@
 
 #include "esp_sleep.h"
 
-void deep_sleep(uint8_t seconds) {
+void deep_sleep(unsigned int seconds) {
     esp_deep_sleep_set_rf_option(2); //
 	esp_deep_sleep(seconds*1000000);	
 }
